Based cppcoro on the implementation of Andreas Buhr to use Linux operations

+++ examples/cppcoro/examples-cc/file_write_read1.cpp
@@ -51,7 +51,7 @@ void mainflow()
 
 		char buffer[1024];
 		char c = 'a';
-		for (int i = 0; i < sizeof(buffer); ++i, c = (c == 'z' ? 'a' : c + 1))
+		for (size_t i = 0; i < sizeof(buffer); ++i, c = (c == 'z' ? 'a' : c + 1))
 		{
 			buffer[i] = c;
 		}
