+++ examples/cppcoro/examples-cl/file_write_read2.cpp
@@ -49,7 +49,7 @@ async_task<void> write(cppcoro::io_service& ioService, std::filesystem::path& fi
 
         char buffer[1024];
         char c = 'a';
-        for (int i = 0; i < sizeof(buffer); ++i, c = (c == 'z' ? 'a' : c + 1))
+        for (size_t i = 0; i < sizeof(buffer); ++i, c = (c == 'z' ? 'a' : c + 1))
         {
             buffer[i] = c;
         }
